Move SOC out of gb-cpu
