+++ gb-cpu/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod address_bus;
 
 pub mod bus;
-pub mod soc;
+pub mod cpu;
 
 pub use address_bus::{AddressBus, Area, FileOperation};
