--- gb-cpu/src/soc.rs
@@ -1,1 +1,0 @@
-pub mod cpu;
